todo fix lag

+++ user_funs.h
@@ -6,4 +6,4 @@ matrix ff0T(matrix, matrix = NAN, matrix = NAN);
 matrix ff0R(matrix, matrix = NAN, matrix = NAN);
 matrix df0(double, matrix, matrix = NAN, matrix = NAN);
 matrix ff1T(matrix x, matrix ud1, matrix ud2);
-int GetFib(int n);
+double GetFib(int n);
